Zombie AI optimization (or is it?)
Store some pathfinding info in a static graph, so it is not recomputed
each time the function is called ; not sure the permformances are
actually better, should benchmark. This can be disabled, anyways.
Zombies can now climb ledges to get off water !
Limited the search of a path to a configurable maximum distance.