Fixed dynamic lighting bugs
The light was not properly initialized on some tiles, causing them to
have an aberrant brightness value, resulting in the whole lighting being
broken.
Moreover, the light updating used to take into account the punctual
lightsources only, not the sunlight.
The game also used to crash when removing a light that was not on the
first layer (it wasn't remove from the lightsources array).
One can now add/remove a tile by clicking, which is useful for testing
purposes.